:construction: back 3 aula

+++ src/routes/memories.ts
@@ -3,8 +3,15 @@ import { z } from 'zod'
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
+    })
     app.get('/memories', async (request, response) => {
+
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub
+            },
             orderBy: {
                 created_at: 'asc'
             }
@@ -28,6 +35,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 id
             }
         })
+        if (!memory.isPublic && memory.userId !== request.user.sub) {
+            return response.status(401).send()
+        }
         return memory
     })
 
@@ -37,19 +47,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
             coverUrl: z.string(),
             isPublic: z.coerce.boolean().default(false)
         })
-        const { content, isPublic, coverUrl} = bodySchema.parse(request.body)
+        const { content, isPublic, coverUrl } = bodySchema.parse(request.body)
         const memory = await prisma.memory.create({
-            data:{
+            data: {
                 content,
                 coverUrl,
                 isPublic,
-                userId: 'a42f758e-35ce-4c9b-81a2-67c4abbe319e'
+                userId: request.user.sub
             }
         })
         return memory
     })
 
     app.put('/memories/:id', async (request, response) => {
+
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -60,13 +71,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
             coverUrl: z.string(),
             isPublic: z.coerce.boolean().default(false)
         })
-        const { content, isPublic, coverUrl} = bodySchema.parse(request.body)
+        const { content, isPublic, coverUrl } = bodySchema.parse(request.body)
 
-        const memory = await prisma.memory.update({
-            where:{
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+        if (memory.userId !== request.user.sub) {
+            return response.status(401).send()
+        }
+
+        memory = await prisma.memory.update({
+            where: {
                 id,
             },
-            data:{
+            data: {
                 content,
                 coverUrl,
                 isPublic
@@ -80,6 +100,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
             id: z.string().uuid(),
         })
         const { id } = paramsSchema.parse(request.params)
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+        if (memory.userId !== request.user.sub) {
+            return response.status(401).send()
+        }
         await prisma.memory.delete({
             where: {
                 id
